Setup Kubernetes for Titanic app
Swagger-codegen made a server with stubbed out endpoints.
The swagger.yml was changed to replace values that caused
naming issues throughout the swagger generated code-base.
The script for running swagger-codegen is included although
it's not intended to be run on a regular basis. In the script,
node.js was chosen for the back-end.

Node.js was chosen for the back-end, and postgres for the db.
Pg-promise is used for convenience, and the db connection
lives in queries.js to separate concerns. Longer SQL
queries have their own sql files under the sql folder.
The schema, as reflected in docker_postgres_init.sql,
shows that
- sex contains enumerated values
- age is a float
- fare is numeric

The mapping of the schema to the expected .json format returned
from the endpoints is eventually enforced in the python tests
run by the github action. Enumerated values turn to string values
without intervention, but float and numeric types need explicit
logic for conversion to integer and float. The controller file assumes the
responsibility for preparing the .json to be sent. Ideally,
the values under the age column from titanic.csv would have
been cleaned before being copied into the postgres table,
as many of them are not integers, but out of convenience,
the responsibility for cleaning the data has leaked into
code inside the controller file.

The db is setup to be seeded when the image is run. There
is no production db connected to the server, meaning changes
made between the image restarting will not be saved. In the
future, it is possible to connect to a production db,
which should be simple as the implementation is located
in one place.

Docker-compose is useful for developing locally, and
in addition, pg-admin is available to examine
changes to the rows. Kompose was used to
convert the docker-compose.yml into kubernetes
manifests. In those manifests, labels were added
for identifying nodes. An ingress using host
api.awesome is part of the manifest, and the
exposed port is 8080.

The scripts used by the GitHub Action to evaluate the code
have been filled and documented. An extra script, setup_local.sh,
is intended to be used to setup the local registry for
docker images to be pushed to and pulled from. The contents
of the script reflect the contents of the GitHub Action so
that the same code can work on the different parts of the CI/CD
pipeline.

With more time, it could be useful to write unit tests,
as that will complement well-structured code.
Observability would be useful as well if this code were to be
deployed.

Currently, all but one of the python tests pass. The swagger.yml
specifies no schema for the delete endpoint, yet the test
checks that the returned .json conform to the PersonData as
defined inside swagger.yml.